LDP-NR/OriginalRecord persistence to Marmotta
The latest version of Marmotta changed how `Content-Type` headers are
handled in POST/PUT requests for creating LDP-RS vs. LDP-NR.
`text/plain` is now assumed to be an RDFSource.

We should find a good way to create/maintain LDP-NR's with content
types like `text/plain` and `application/xml`, for which Marmotta will
try to create an LDP-RS. This may be an interaction model thing.

In the meanwhile, OriginalRecords are created with a Content-Type of
`application/octet-stream`.