Use the new DPLA marmotta-jetty
This unifies us on a single servelet container for Solr and Marmotta.
It also pins us to a snapshot of Marmotta 3.3.0, which current LDP
work depends on. We're expecting an official 3.3.0 release imminently.

When building for the first time with this update, you may need to
remove any lingering Marmotta files by doing `rm -rf /tmp/marmotta`.y