Add Enrichment and FieldEnrichment interfaces
Enrichments:
------------

Sets the values from some number of specified input fields to some
number of specified output fields after processing them with the
private interface method `#enrich_value`.

Fields are selected by passing hashes to the `input_fields` and
`output_fields` parameters of `enrich`. An example hash might be:

    {:sourceResource => {:creator => :name}}

Field Enrichments:
-------------------

This is a narrower case of `Krikri::Enrichment` which runs the
enrichment against each of the specified fields in turn, setting the
field's value to the result.