Pessimistically constrain sass-rails to match Rails 4.1 generator
Since sass-rails 5.0.0 was released, Krikri's Bundler installation was using it by
default, despite the Rails 4.1 generator using ~> 4.0.3 as a pessimistic constraint.
This caused a conflict when generating a new app, such as the one created using
engine_cart to run tests in CI.

Also addresses a minor change in a comment added to the Gemfile created by the
generator.