Add a registry of harvesters
- Turn Krikri::Harvester into a mixin module
- Create Krikri::Registry and have Krikri::Harvester::Registry
  and Krikri::Mapper::Registry extend it.

The registry will be used when we enqueue jobs, or in the eventual user
interface, in order to retrieve a harvester class given a token (e.g.
get an OAIHarvester class for "OAI") or to retrieve a list of available
harvester tokens for the purpose building menus.

A future Event will look up the harvester token from the HarvestSource
for a given institution, and call #enqueue on the relevant Harvester to
put it in the queue.