Add discovery of Harvester options
Krikri::Harvester::expected_opts has been added to help client code know
what options to get from the user to store in HarvestSource. See the
comment in that method.

An opts field has been added to HarvestSource for storing serialized
options that get passed to the harvester. For example, a list of sets
for an OAI harvester.  This serves to namespace options that are
harvester-specific.

OAIHarvester has also been changed to take 'uri' as an option instead of
'endpoint', to bring it inline with the HarvestSource model, which has a
'uri' field, not an 'endpoint' field.