Allow force-registering of Registry items
- Allow Registry to force-register items by adding #register! in
  addition to existing #register
- Update Krikri::Mapper#define to use Registry#register!

This takes care of an issue wherein Resque's resque:work rake task
calls Rails.application.eager_load!, which causes Mapper#define to get
called twice, for reasons that aren't clear.  When we analyzed the
situation, we realized that it was probably overzealous for
Registry's only registration method to throw an exception if the name
was already registered.  Registry#register! has been added to provide an
alternative.