Extend Mapping DSL with record access
Adds functionality to the mapping DSL for creating instances of a
resource class from each of a given parser node's values and for
accessing parsed values in a more human readable fashion.

Creates a RecordProxy in the mapping DSL context. The special DSL method
chain, with method names, their arguments (and blocks to pass through)
which will be called against the parsed record (via ValueArray) when the
record is processed. RecordProxy mimics Proc in that it has #call and
same.

Adds a ValueArray, which encapsulates common access patterns for parsed
record nodes. ValueArray's methods return instances of itself, as they
intended to be chained. E.g.

```ruby
val.field(:creator).select { |v| v.some_attribute == 'attribute value' }
```

ValueArray should be extended over time to wrap more of Array's
functionality and capture any emerging patterns.

Lastly, this enhances ChildDeclaration to accept optional :each and :as
options. These options specify that an instance of the requested child
class will be created for each value in the Proc (or RecordProxy) passed
to :each, and that the matching values will be accessible as a method
registered in the block's scope with the name passed to :as.

```ruby
Krikri::Mapper.define(:esdn_mods) do
  sourceResource :class => DPLA::MAP::SourceResource do
    contributor :class => DPLA::MAP::Agent, :each => record.field('mods:name').select { |name| name['mods:role'].map(&:value).include?('contributor') }, :as => :contrib do
      providedLabel contrib.field('mods:namePart')
    end

    creator :class => DPLA::MAP::Agent, :each => record.field('mods:name').select { |name| name['mods:role'].map(&:value).include?('creator') }, :as => :creator_role do
      providedLabel creator_role.field('mods:namePart')
    end
 end
end
```