Add date normalization
Implements two date enrichments:

ParseDate converts string values in individual fields to Date objects.
This enrichment is meant to be run on TimeSpan attributes #begin/#end
when they have been mapped directly.

TimespanSplit accepts a string value or a TimeSpan, and returns a
timespan object with a single #begin and #end, if any can be found. If
no #begin and #end are set, they are parsed from #providedLabel.

Both rely on Krikri::Util::ExtendedDateParser, which parses strings to
one of Date, EDTF::Interval, or EDTF::Epoch. The last two represent
date ranges; when they are returned to TimespanSplit it uses their
methods to determine the start and end of the range.