DSL: add ValueArray#first and improve empty nodes
Creates a `#first_value` method in the `Krikri::Parser::ValueArray`. The
new method works like `Array#first`, but wraps the value in a
`ValueArray` so it can be chained like other parser methods.

Also cleans up the handling of empty nodes to uniformly return an empty
`ValueArray`. Adds tests to this effect in the shared Parser examples.