Refactor Krikri::Harvester::CouchdbHarvester using Analysand
This is an alternative implemenation of a harvester from CouchDB views
using Analysand (https://github.com/yipdw/analysand). Analysand supports
streaming view responses and uses JSON::Stream for parsing.

The gemspec file refers to a DPLA pre-release, which is synced to HEAD
on the upstream master (since the release was not actually issued
and published to Rubygems).