Add parser arguments to mapping definitions
Krikri::Mapper::Agent receives only unparsed records, assuming that the
Mapping will handle parsing through `#process_record`.  The parser to
use is specified through the mapping definition with the `:parser`
option.

Previously, there was no way to pass arguments to the parser. Arguments
are not needed in most cases, but necessary for some `Krikri::XmlParser`
subclasses.  This adds `:parser_args`, allowing you to pass an array
of parser arguments when defining a mapping.

    Krikri::Mapper.define(:scdl_qdc,
		          :parser => Krikri::QdcParser,
    			  :parser_args => ['//qdc:qualifieddc']) do
      ...
    end