Simplify `#enqueue` interface and dependencies
Moves `#run` logic from `Krikri::Job` subclasses into the base class.
All job subclasses currently assume that `SoftwareAgent` implementers
respond to `#run(activity_id)`, this pushed that logic up to `Job`,
removing the duplication.

With that done, the only remaining responsibility of the Job subclasses
is to specify a queue name. References to the class name constants from
`SoftwareAgent` become proxies to simple strings that represent the
queue_names. The introduction of `SoftwareAgent#queue_name` removes the
obfuscation

`SoftwareAgent#enqueue` is simplified to accept an options hash alone
(still defaulting to {}), with an optional queue name as the first
argument.