Set fragment URI for first sourceResource on ore:Aggregations
In MAPv3.1 data, we have 1 and only 1 dpla:SourceResource associated with
each ore:Aggregation, and by default, it gets assigned the subject
`http://base_uri/items/id#sourceResource`. This modifies the initializer
in Krikri to add the same behavior.

If we end up having more than one dpla:SourceResource on a given
ore:Aggregation, we will probably want to deal with this in a cleaner way.

Thanks to @no-reply for the improved tests.