Add and refactor search indexing, and add behaviors
- Add behaviors (in `entity_behaviors`) to provide different answers to
  generated-entities questions depending on the type of activity.
- Add `EntityConsumer` mixin for classes, like certain software agents,
  that generate entities.
- Update `Krikri::Mapper::Agent` because of the changes above
- Replace `IndexService` with `SearchIndex` and its subclasses,
  `QASearchIndex` and `ProdSearchIndex`
- Add `Activity#entitiy_uris`, which enumerates URI strings of entities
  that pertain to the activity.
- Add `Activity#entities`, which returns entities (for example,
  aggregations) that were affected by the activity. A behavior's
  `#entities` method is invoked for this purpose.
- Add the ElasticSearch service to the development VM. Add tests that
  are not runnable in Travis CI, but that work in the VM with
  ElasticSearch 0.90 (which can't be installed in a Travis container).

The buildout of search indexing has required us to think about how we
can treat provenance consistently for different kinds of activities. The
new behavior classes and the additions of the Activity model make it
possible to ask for the entities that were affected by an activity,
using consistent interfaces. A mapping activity may have consumed the
original records generated by a harvest, but an enrichment and an
indexing activity are usually going to consume aggregations, when KriKri
is used within the DPLA context. In the case of another organization,
it's possible that the entities generated by mappings and enrichments
could be something else, thus the level of abstraction that you see
here.