Factor aggregation-specific logic out of Activity
Activity should be able to produce lists of generated entity URIs,
but should not assume that the generated entities are aggregations.
To this end, move such aggregation-specific code to the SearchIndex,
where it is really needed.

This will probably get another refactor soon, as we implement this
behavior with Mapper, which needs to retrieve generated original
record entities for an activity.  We'll see then what code is at risk of
duplication.