Adds QA Reports MVC
Creates a QAReports model that consumes the QA query client, to create
provider-based reports (as hashes), and allows them to be serialized as
CSV. These are ActiveRecord objects, stored in the database.

Provides controllers and views for the QAReports, allowing CSV downloads.
This does not establish a pattern for generating new reports.

Also adds a significant number of tests for existing views,
establishing the beginnings of a practice of using `rspec-rails` view
specs. Note that helpers are stubbed, due in part to an issue with
engine routes in `rspec-rails`. This suggests that we should consider
helper specs as contract tests for helpers our views rely on.