Krikri::Provider#reload query should use a DISTINCT clause
Krikri::Provider#reload builds out the triples associated with the
provider through a SPARQL query. When there are more than a small
number (say, 100) of mapped records in your repository, it takes
a very long time for it to build out the provider given that #reload
calls set_value multiple times: one for each of the sum of the
number of solutions for each mapped record associated with that
provider.