Remove record counts from provider index view
The provider index view ("dashboard") calls the SPARQL query
associated with Krikri::Provider#records for each provider listed.
The counts are definitely useful, but the performance issues we're
seeing now are preventing the page loads from completing currently.
It may be more useful to generate the counts from Solr index, or
figuring out another way of caching them for each activity (e.g.
in the database).