Do not obliterate defaults, QASearchIndex constructor
Have the QASearchIndex constructor preserve default configuration
properties that are not being passed in via the `opts` argument to
`#initialize`.

This allows the Solr URI in Krikri::Settings.solr to be passed to
`RSolr.connect`.

Fixes https://www.pivotaltracker.com/story/show/91822774

Also add a regression test for Krikri::QASearchIndex#initialize that
verifies that default options are preserved when extra ones are given.