Forceably require sass-rails ~> 5.0.0
Krikri currently is pinned to the Rails 4.1 release series, which depends
on sass-rails ~> 4.0.3. Recently, we ran into an issue which seemed to be a
manifestation of https://github.com/sass/sass/issues/1656 despite the seeming
version differences.

The changes that this makes are as follows:

* Explicitly add sass-rails ~> 5.0.0 to the Krikri gemspec. Without this, we
run into a dependency conflict with tilt (sass-rails requires a different version
than haml does, which is required by rdf-rdfa through the linkeddata gem).
* Pass `--skip-sprockets` to the engine_cart handler that generates the new
application. If this is not included, you get some annoying messages about
dependency conflicts that shouldn't be show stoppers.
* Add the necessary sass-rails and uglifier gems, and manually re-enable
the sprockets railtie.