Pin bootstrap-sass to 3.3.4.1 (FIXME/rails 4.2)
This relates to sass/sass#1656. bootstrap-sass 3.3.5 introduced a
regression which causes the Blacklight default CSS to fail its build.
A quick fix could be to require sass-rails 5.0.x, but that introduces
some slight dependency resolution problems under Rails 4.1 since
Rails 4.1 apps use sass-rails ~> 4.0.3 by default. The better fix
will be to upgrade Krikri to Rails 4.2, but that will require further
testing. This change has to be in the gemspec, and not the Gemfile for
Krikri or the application in which it is hosted, as the dependencies
won't resolve properly otherwise.