Error handling in Mapping
Introduces a `Mapping::Error` that tracks multiple errors for a given
record and handles messaging.

In the `Mapping`, we catch errors for each declaration on a given
record, storing them with the relevant property name, and re-raise. The
`Mapper` then simply catches the error and logs its message.

This simplifies handling outside the DSL code, and creates richer error
messages.