Pin to RDF 1.1.x
RDF 1.99.0 was released recently and introduces many deprecation
warnings in preparation for the RDF 2.0 release coming around the turn
of the calendar year. We expect 1.99.1, and perhaps some other minor
releases between now and then. 1.99.1 will contain most or all of the
features that make it into 2.0.

This locks us to `1.1.x` until we can make a deliberate upgrade. My
suggestion would be to wait for that until *at least* 1.99.1, reducing
the number and complexity of the upgrades.