Allow OAI to recover on errors when iterating sets
Adds error recovery when an attempt to retrieve a set throws an
`OAI::Excepton`. OAI-PMH requests throw errors when no matches can be
found with (e.g.) the required metadata prefix. When this happens we
want to pass over the set, logging & returning empty, instead of
aborting the harvest.

Note that the failover applies to *any* OAI exception. We've observed
that some provider endpoints report different errors, and
`OAI::Exception` does little to help us identify error types. This may
turn out to be heavy-handed.