Fix failed ProdSearchIndex spec that was skipped in CI
This reorders the spec to used the :opts passed by RSpec to be only used in
the context describing `#initialize`, and removes the mock that was preventing
the appropriate method from firing. It also modifies the `entities query`
shared context to set subject URI for the built aggregations.

The specs that test `Krikri::ProdSearchIndex` with an Elasticsearch connection
were being skipped in Travis, and they failed for me locally. It looked in
some cases as though it was trying to make a connection to
http://example.org/moomin-index/moomin as used in the describe block.

Adds tests that the search index actually responds with the correct ids when
running `#update_from_activity` (HT: @no-reply).