Refactor `AsyncUriGetter` to cache `Faraday` conn
Caches the `Faraday` connection when first accessing `Request#http` for
the first time. This shouldn't have a tangible impact with the code as
written, but is more resilient in the event that `Request` ends up with
more than one call to `#http` at a future date.

Contains some YARD documentation & style updates are also present.