Limit `RDF::URI.cache` capacity
By default, `RDF::URI::CACHE_SIZE` is `-1` and all URIs initialized with
`RDF::URI.intern` are cached for the life of the application. In job
scenarios, this leads to a slow growth of the cache, as more unique URIs
are added. For a good number of the URIs we're using (e.g. LDP URIs and
their `#sourceResource` fragments), use is scoped to the processing of
one record. Limiting the cache size allows these to cycle out, while
commonly used URIs should stay in the cache avoiding duplicate objects
in memory.

The cache size is given by `Krikri::Settings['uri_cache_size']`, and
defaults to `1_000_000` (cached URIs). We could choose to change the
cache size on various worker boxes by installing unique settings
values.

Setting the cache size at application startup causes a warning to be
thrown that the previous value of the "constant" `CACHE_SIZE` is being
overwritten. An issue is open to find a way to squash the warnings,
hopefully with a better cache capacity management pattern, at
https://github.com/ruby-rdf/rdf/issues/256