Allow OAI harvesters to use alternate ID sources
By default, we use the `identifier` field of the OAI header to seed the
record id minter. In the legacy system, the ids used were derived from
the `dc:identifier` field in `metadata`.

This patch allows harvesters for those sources to select a field with an
xpath expression, passed as `[:oai][:id_path]` in options.