Raise error for empty original record, minting ID
Raise an exception when minting an ID in DPLA::MAP::Aggregation#mint_id!
if there is no original record, instead of generating a random one.

As previously written, if a mapping failed to map originalRecord, you
would end up with two RDFSources, one for the harvested record, and one
for the mapped record with a random ID.

Also improve the existing error messaging for
DPLA::MAP::Aggregation#original_record when there's no original record.