Fix harvester_spec.rb test with wrong expect subject
In Harvester spec, fix ".entity_behavior knows its entity_behavior" to
expect `subject.class.entity_behavior` instead of
`klass.entity_behavior`.

The latter one should fail, because `klass` should not be expected
to have `include`d `Krikri::Harvester` yet.  Depending on the order in
which the specs were run, the issue was being masked, because the spec
was operating on a `klass` that had had `Harvester` included by an
earlier spec.