Adds provenance count queries [DO NOT MERGE]
Adds support for `ProvenanceQueryClient#count_by_activity` using
SPARQL's `COUNT` functionality. Refactors `ProvenanceQueryClient`
significantly in the process.

`Krikri::Activity` gets an `#entity_count` method, which is much faster
in practice than `#entity_uris.count`. This allows optimized counting on
the server side, instead of returning an entire result set and iterating
through the `RDF::Query::Solutions`.

Note: this commit is WIP. The pending test on `#count_by_activity` is
critical, but fails on H2 when the `|` operator is used in the
predicate. This is being pushed up to track the WIP, but needs testing
in a more production like environment before it can be merged. It may be
that we'll need to report a bug on Marmotta and leave the test marked as
pending, for the moment.