Remove Solr docs more carefully in test suite
The particular issue was in the specs for
`Krikri::QASearchIndex#update_from_activity`. These got left behind,
some tests had apparently compensated for this by cleaning up in
`before`. That behavior is removed, to make it easier to catch issues
like this in the future.

Also normalizes solr index cleanup on the `clear_search_index` helper. I
think this was added after some of these specs were written, and some of
them never got migrated.