Restores `ValueArray` in scoped DSL variables
"Variables" declared as scoped DSL methods were mistakenly cast to their
`#value` when we want a `ValueArray`. This restores the old behavior and
adds an integration test.

This could still use a significant refactor. There is probably a way to
do this without nested iterations through the values. Additionally,
`RecordProxy#dup` is customized in a way that makes this somewhat
tightly coupled and hard to read.

Fixes https://issues.dp.la/issues/8522.