Allow Travis failures for unsupported Rubies
Krikri is still in version 0.x.  DPLA develops with Ruby version 2.1.3
and is still deploying with that version. Bundler 1.13.0 was recently
released, and it generates errors evaluating our Gemfile. The
known-good version of bundler, 1.11, can be installed with
before_install in .travis.ci, but its use is overridden when Travis
runs -- `bundle install' happens with 1.13 and I haven't figured out how
to change that. This does not happen with the 2.1.3 build.

Because of time pressures, I am moving Ruby versions 2.1.6 and 2.2.3
into the `allow_failures' list in .travis.yml. I intend to revisit the
issue with Travis and fix the document the root cause of the bundler
failures.

Note that the existing `before_install' directive in .travis.yml has not
been effectively enforcing "~> 1.11.0" even for successful builds. (Check
the Travis build logs, e.g.
https://travis-ci.org/dpla/KriKri/jobs/158281268 .)