Remove existing Bundler 1.13.0 installs on Travis (fixes #8556)
Dependency resolution in Bundler 1.13 breaks when using engine_cart, which we use
in Krikri to build a test application for the engine. The changes to the Travis
build are modified from the workaround specified in cbeer/engine_cart#64, but
the better long term fix would be to upgrade engine_cart to v1.0.1 (see #8557),
and to ensure Krikri works with the latest Bundler.